Create Excel Macro.vb
A couple of the macros that I have used during work for excel. The first macro is for adding a timestamp in a desired column whenever you add data into a corresponding column. The second macro will make a 'macro-free' duplicate of the excel file that you are working with once you close the file. It will add today's date to the beginning of the filename.